refactor: 말과 차 공통부모 정의
